Switch networking from unity's UNET to Photon's PUN2
Unity's UNET was extremely limited and incomplete when it came to networking. I switched over for perforamce, and to support a lot of vital features that UNET lacked (Game room options, search filtering, etc).

The networking state is exactly where it was with the previous UNET system...mostly there, but not quite complete (players, bullets, and turret look direction are snyc'd).